Shift Lights working

+++ Core/Inc/shiftLights.h
@@ -18,18 +18,18 @@
 void LightsStartUpSequence();
 
 void setColor(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t green, uint8_t red,
-		uint8_t blue, uint8_t *ledcolors, uint16_t *ledbytes, int ledIndex);
+		uint8_t blue, uint8_t *ledcolors, uint32_t *ledbytes, int ledIndex);
 
-void shiftLightsInit(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t *ledcolors, uint16_t *ledbytes);
+void shiftLightsInit(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t *ledcolors, uint32_t *ledbytes);
 
 void HAL_TIM_PWM_PulseFinishedCallback(TIM_HandleTypeDef *htim);
 
 void UpdateShiftLights(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t *ledcolors,
-		uint16_t *ledbytes, int current_rpm, int * RPM_thresholds);
+		uint32_t *ledbytes, int current_rpm, int * RPM_thresholds);
 
 void lightOn(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t *ledcolors,
-		uint16_t *ledbytes,int index);
+		uint32_t *ledbytes,int index);
 
-void startUp(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t *ledcolors, uint16_t *ledbytes);
+void startUp(TIM_HandleTypeDef *htim, uint32_t Channel, uint8_t *ledcolors, uint32_t *ledbytes);
 
 #endif
